@@ -1,9 +1,8 @@
 package com.charts.api.ticket.enums;
 
 import com.charts.general.entity.enums.EnumUtils;
-
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
 public class TicketTypeConverter implements AttributeConverter<TicketType, String> {
