@@ -2,9 +2,8 @@ package com.charts.general.entity.enums.converters;
 
 import com.charts.general.entity.enums.EnumUtils;
 import com.charts.general.entity.enums.types.Months;
-
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
 public class MonthConverter implements AttributeConverter<Months, String> {
